+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0 (Jun 25, 2025)
+
+- add "Show first letter" button to lessons
+- add average correctness percentage for each vocab
+- vocab words are now sorted
+- fix a bug when special symbols were added to the end of the string instead of where the cursor was
+
+
 ## 1.0.0 (Apr 10, 2025)
 
 - Initial release.